server/requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.34.2
python-multipart==0.0.20

# LlamaIndex core + integrations
llama-index-core==0.12.29
llama-index-llms-ollama==0.5.0
llama-index-embeddings-huggingface==0.5.2
llama-index-vector-stores-chroma==0.4.1
llama-index-readers-file==0.4.7

# Embedding model
sentence-transformers>=3.0.0

# Vector store
chromadb==1.0.7

# PDF parsing
pypdf==5.5.0

# HTTP client
httpx==0.28.1
